EEPROM 25AA1024 write, read 100% working

+++ Core/cmd_dispatcher/cmd_dispatcher.h
@@ -1,6 +1,6 @@
 #ifndef CMD_DISPATCHER_H_INCLUDED
 #define CMD_DISPATCHER_H_INCLUDED
 
-uint8_t CmdDispatch(const uint8_t* const pStrCmd, const uint8_t lng);
+uint8_t CmdDispatch(const uint8_t* const pStrCmd, const uint16_t lng);
 
 #endif // CMD_DISPATCHER_H_INCLUDED
